backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12

# Cache
redis==5.2.1

# Payments
razorpay==1.4.1

# Maps
googlemaps==4.10.0

# HTTP client
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
python-dateutil==2.9.0

# Dev only
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
